+++ src/renderer/components/atoms/text/helper/build-styles-for-document.ts
