+++ src/renderer/components/atoms/text/helper/build-styles-for-interface.ts
