+++ src/renderer/components/atoms/text/helper/build-styles-for.ts
@@ -1,7 +1,7 @@
 import type { TextUsage, TextVariant } from '../types';
 import { DOCUMENT_VARIANTS, INTERFACE_VARIANTS } from './constants';
-import buildStylesForDocument from './buildStylesForDocument';
-import buildStylesForInterface from './buildStylesForInterface';
+import buildStylesForDocument from './build-styles-for-document';
+import buildStylesForInterface from './build-styles-for-interface';
 
 const buildStylesFor = (variantName: TextVariant, as: TextUsage) => {
   if (DOCUMENT_VARIANTS.includes(as)) {
