+++ src/renderer/components/atoms/text/helper/build-text-color-for-document.ts
