+++ src/renderer/components/atoms/text/helper/index.ts
@@ -1,3 +1,3 @@
-export { default as buildStylesFor } from './buildStylesFor';
-export { default as returnTextColor } from './returnTextColor';
-export { default as resolveElement } from './resolveElement';
+export { default as buildStylesFor } from './build-styles-for';
+export { default as returnTextColor } from './return-text-color';
+export { default as resolveElement } from './resolve-element';
