+++ src/renderer/components/atoms/text/helper/resolve-element.ts
